Move Logger to helpers
